feat(47): improved file list components structure for delete file functionality
