+++ src/components/Layouts/MainLayout/FilesPanel/types.ts
@@ -1,0 +1,5 @@
+import { IFileInfo } from '../../../../types/file/fileTypes';
+
+export interface IFileItemProps {
+  file: IFileInfo;
+}
